Add Real Dataset (#1)
* Add a script for adding the real dataset.
* Fix a database table type error (video scale should be FLOAT, not INT).
* Add error messages for the client initially if the server is down.
* Add a script for adding videos on the background.

+++ back-end/www/application.py
@@ -106,7 +106,7 @@ class Video(db.Model):
     # For example, a point (x, y) on the video has a global coordinate (left + x*scale, top + y*scale)
     width = db.Column(db.Integer, nullable=False)
     height = db.Column(db.Integer, nullable=False)
-    scale = db.Column(db.Integer, nullable=False)
+    scale = db.Column(db.Float, nullable=False)
     # For the global large panorama, the origin (0, 0) is at the top-right corner
     # The top-left point coordinate (x, y) of the video relative to the large panorama
     left = db.Column(db.Integer, nullable=False)
@@ -355,6 +355,7 @@ def get_pos_labels():
         q = Video.query.filter(Video.label_state.in_((0b10111, 0b1111, 0b10011))).paginate(page_number, page_size, False)
     else:
         q = Label.query.filter(Label.user_id==user_id).from_self(Video).join(Video).filter(Video.label_state.in_((0b10111, 0b1111, 0b10011))).distinct().paginate(page_number, page_size, False)
+        #q = Label.query.filter(and_(Label.user_id==user_id, Label.label==1)).from_self(Video).join(Video).distinct().paginate(page_number, page_size, False)
     return jsonify_videos(q.items, total=q.total)
 
 """
